io: move poll defines to internal/io.h
USE_POLL is defined in io.c and thread.c with different semantic. Move
to internal/io.h and adjust the #ifdef to #if.

Also move POLL*_SETs. Define POLLERR_SET on Linux as well. This
genereates a little more code in rb_wait_for_single_fd compated to teh
case when it was defined to 0 but should not impact result. We need
POLLERR_SET defined for checking errors.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>